css/selectors/invalidation/is.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Selectors Invalidation: :is()</title>
    <link rel="author" title="Victoria Su" href="mailto:victoriaytsu@google.com">
    <link rel="help" href="https://drafts.csswg.org/selectors-4/#zero-matches">
    <meta name="assert" content="This tests that the :is() selector is effective">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      .b {
        color: yellow;
      }
      /*Simple selector arguments */
      :is(.b, .c) {
        color: red;
      }
      /*Compound selector arguments */
      .a~:is(.c#d, .e) {
        color: green;
      }
      /* Complex selector arguments */
      .h {
        color: red;
      }
      :is(.a~.h, .a~.h+.f) {
        color: yellow;
      }
      /* Nested */
      :is(.a>:is(.g+.h, .b)~.i) {
        color: blue;
      }
      :is(:matches(.a~.h)) {
        color: yellow;
      }
    </style>
  </head>
  <body>
    <div id="a1">
      <div class="g">
      </div>
      <div class="h">
      </div>
      <div class="i" id="i1">
        Blue
      </div>
    </div>
    <div class="b" id="b1">
      Yellow
    </div>
    <div class="c" id="c1">
      Red
    </div>
    <div class="c" id="d">
      Green
    </div>
    <div class="h" id="h1">
      Red
    </div>
    <div class="f" id="f1">
      Yellow
    </div>
    <script>
      document.body.offsetTop;

      var black = "rgb(0, 0, 0)";
      var green = "rgb(0, 128, 0)";
      var red = "rgb(255, 0, 0)";
      var yellow = "rgb(255, 255, 0)";
      var blue = "rgb(0, 0, 255)";

      test(() => {
        assert_equals(getComputedStyle(b1).color, yellow);
        assert_equals(getComputedStyle(c1).color, red);
        assert_equals(getComputedStyle(d).color, red);
        assert_equals(getComputedStyle(h1).color, red);
        assert_equals(getComputedStyle(f1).color, black);
        assert_equals(getComputedStyle(i1).color, black);
      }, "Preconditions.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(b1).color, yellow);
        assert_equals(getComputedStyle(c1).color, red);
      }, "Invalidate :is() for simple selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(d).color, green);
      }, "Invalidate :is() for compound selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(h1).color, red);
        assert_equals(getComputedStyle(f1).color, yellow);
      }, "Invalidate :is() for complex selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(i1).color, blue);
      }, "Invalidate nested :is().");

    </script>
  </body>
</html>
